Create README.md
If copied into the new Advanced Formula Environment in Excel this file will add a function Allocateλ which will take a set of input volumes (purchases) and allocate them to a set of output demands (sales), producing an array to show which input batches were used to meet what order.  If purchase costs are provided the array will aggregated to show the input costs associated with each output.
Other uses of the formula include evaluation the effect a manufacturing cap (a single input) has on the ability to deliver a series of orders (met in full being followed by part fulfilment and, finally, unfulfilled.  Equally the function will allocate income to tax bands provided the tax bands are defined by width, allowing the formula to evaluate thresholds.
Supporting functions are Accumulateλ, which generates running totals for a list of values, as well as Sumλ(x) and Addλ(x,y) which take their parameters from Lambda helper functions.